src/deployment/inference/requirements.txt
Flask
gunicorn
google-cloud-storage
tensorflow
Pillow
numpy
SimpleITK
scikit-image
opencv-python-headless
